Don't block the message loop on Docker builds
This swaps out subprocess.Popen and communicate calls, which
block, when invoking the Docker CLI, and instead uses
asyncio.create_subprocess_exec, which offers equivalent functionlity
without blocking the message loop.

This addresses the responsiveness issues and behaves better with
the Pulumi CLI. In particular, for lengthy builds, we no longer
interfere with the ability of the CLI to report progress and display
resources as they are created by the program, asynchronously
interleaved amongst the build itself.

This is technically a breaking change, however, because the
functions now return futures where they used to return raw values.
It would be reasonable to, instead of changing these in place,
turn them into _async routines (which are used internally), and
then have thin wrappers that block alongside them.

This change fixes #313.